Add files via upload
This is the data collected from Area X.O using Licor Sensor.  For 2021 data, the chamber (senor in a chamber) is placed on the patch of the field that has broadcast fertilizer applied and worked into the top 3” of soil (Conventional Broadcasting) (no crop), 225kg/ha of urea and polymer-coated urea fertilizer was applied. For the 2022 data, the chamber is placed on the strip/row using Conventional Broadcasting (without crop), 225kg/ha of urea and polymer-coated urea fertilizer was applied.